requirements.txt
mutagen>=1.47.0
python-discogs-client>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.1
